+++ src/main/java/org/example/repository/QueryGeneratorFactory.java
@@ -2,4 +2,5 @@ package org.example.repository;
 
 public interface QueryGeneratorFactory {
     String createTableQuery(Class<?> clazz);
+    // crud
 }
