+++ src/main/java/org/example/service/PostgresDatabaseService.java
@@ -1,51 +1,10 @@
 package org.example.service;
+
 import java.sql.Connection;
-import java.sql.Statement;
-import org.example.repository.PostgresQueryGeneratorConcrete;
 
-public class PostgresDatabaseService implements DatabaseService {
-    private Connection connection;
-    private PostgresQueryGeneratorConcrete queryGenerator = new PostgresQueryGeneratorConcrete();
-
-    public PostgresDatabaseService(Connection connection) {
-        this.connection = connection;
-    }
-
+public class PostgresDatabaseService extends DatabaseServiceConcrete {
     @Override
-    public void createTable(Class<?> entity) {
-        String createTableQuery = queryGenerator.createTableQuery(entity);
-        try (Statement stmt = connection.createStatement()) {
-            stmt.executeUpdate(createTableQuery);
-            System.out.println("Table created successfully");
-        } catch (Exception e) {
-            System.out.println(e);
-        }
-    }
-
-    @Override
-    public void insert(Object entity) {
-
-    }
-
-    @Override
-    public void update(Object entity) {
-
-    }
-
-    @Override
-    public void delete(Object entity) {
-
-    }
-
-    @Override
-    public void closeConnection (Connection con) {
-        try {
-            if(con != null && !con.isClosed()) {
-                con.close();
-                System.out.println("Close connection success");
-            }
-        } catch (Exception e) {
-            System.out.println(e);
-        }
+    public DatabaseService createDatabaseService(Connection connection) {
+        return new PostgresDatabaseServiceConcrete(connection);
     }
 }
